test: step3 테스트 클래스 분리
